Updated code with working resolve page
Resolve page works now so that you can click on the marker you would like to remove and it removes it from the page. The report and resolve pages have also been made so that you have to be logged in in order to use them (CHECK IF THIS IS CORRECT).